@@ -2,7 +2,7 @@
   "name": "ani-bot",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "index.ts",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "ts-node src/index.ts",
@@ -11,35 +11,23 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@discordjs/builders": "^0.4.0",
-    "@discordjs/rest": "^0.1.0-canary.0",
-    "@types/express": "^4.17.13",
+    "@types/express": "^4.17.21",
     "@types/filesystem": "^0.0.32",
-    "@types/turndown": "^5.0.1",
-    "axios": "^0.21.1",
-    "crypto-js": "^4.1.1",
-    "discord-api-types": "^0.22.0",
-    "discord.js": "^13.1.0",
+    "@types/turndown": "^5.0.4",
+    "axios": "^1.6.8",
+    "crypto-js": "^4.2.0",
+    "discord.js": "^14.14.1",
     "dotenv": "^10.0.0",
-    "express": "^4.17.1",
+    "express": "^4.19.2",
     "fs": "^0.0.1-security",
-    "graphql": "^15.5.1",
-    "graphql-request": "^3.4.0",
-    "mongoose": "^5.13.7",
-    "node-fetch": "^2.6.1",
+    "graphql": "^15.8.0",
+    "graphql-request": "^3.7.0",
+    "mongoose": "^5.13.22",
+    "node-fetch": "^2.7.0",
     "path": "^0.12.7",
-    "pm2": "^5.1.1",
-    "ts-node": "^10.2.1",
-    "turndown": "^7.1.1",
-    "typescript": "^4.3.5"
-  },
-  "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^4.29.2",
-    "@typescript-eslint/parser": "^4.29.2",
-    "eslint": "^7.32.0",
-    "eslint-config-standard": "^16.0.3",
-    "eslint-plugin-import": "^2.24.1",
-    "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^5.1.0"
+    "pm2": "^5.3.1",
+    "ts-node": "^10.9.2",
+    "turndown": "^7.1.3",
+    "typescript": "^5.4.5"
   }
 }
