@@ -1,7 +1,6 @@
-import { CommandInteraction, SelectMenuInteraction } from 'discord.js'
+import { ActionRowBuilder, CommandInteraction, ComponentType, StringSelectMenuBuilder } from 'discord.js'
 import { Aliases, AniMedia, Parties } from '../../../types'
 
-const Discord = require('discord.js')
 const wait = require('util').promisify(setTimeout)
 const { request } = require('graphql-request')
 const { GET_MEDIA } = require('../../../queries')
@@ -11,7 +10,7 @@ module.exports = {
     name: 'leave'
   },
   async execute (interaction: CommandInteraction, thisServerParty: Parties, serverAliases: Aliases, serverExists: boolean) {
-    const id = `<@!${interaction.user.id}>`
+    const id = `<@${interaction.user.id}>`
     if (serverExists) {
       const userList = serverAliases.server.users
       const user = userList.find(x => x.userId === id)
@@ -27,9 +26,9 @@ module.exports = {
             titles.push(addToTitles)
           })
 
-          const row = new Discord.MessageActionRow()
+          const row = new ActionRowBuilder<StringSelectMenuBuilder>()
             .addComponents(
-              new Discord.MessageSelectMenu()
+              new StringSelectMenuBuilder()
                 .setCustomId('select')
                 .setPlaceholder('Nothing selected')
             )
@@ -44,7 +43,7 @@ module.exports = {
 
           await interaction.reply({ content: 'Select a watch-party to leave:', components: [row], ephemeral: true })
 
-          const filter = async (i: SelectMenuInteraction) => {
+          const filter = async (i: any) => {
             if (i.user.id === interaction.user.id) {
               const titleToLeave = i.values[0]
               const anime: AniMedia = await request('https://graphql.anilist.co', GET_MEDIA, { search: titleToLeave })
@@ -61,7 +60,7 @@ module.exports = {
             }
             return i.user.id === interaction.user.id
           }
-          interaction.channel!.awaitMessageComponent({ filter, componentType: 'SELECT_MENU', time: 15000 })
+          interaction.channel!.awaitMessageComponent({ filter, componentType: ComponentType.StringSelect, time: 15000 })
         }
       } else {
         interaction.reply({ content: 'You have not yet been aliased to an AniList user. `/alias add`', ephemeral: true })
