@@ -1,8 +1,7 @@
-import { CommandInteraction, SelectMenuInteraction } from 'discord.js'
+import { ActionRowBuilder, CommandInteraction, ComponentType, StringSelectMenuBuilder } from 'discord.js'
 import { AniMedia, Parties } from '../../../types'
 
 const { request } = require('graphql-request')
-const Discord = require('discord.js')
 const wait = require('util').promisify(setTimeout)
 const { GET_MEDIA } = require('../../../queries')
 
@@ -21,9 +20,9 @@ module.exports = {
         titles.push(addToTitles)
       })
 
-      const row = new Discord.MessageActionRow()
+      const row = new ActionRowBuilder<StringSelectMenuBuilder>()
         .addComponents(
-          new Discord.MessageSelectMenu()
+          new StringSelectMenuBuilder()
             .setCustomId('select')
             .setPlaceholder('Nothing selected')
         )
@@ -38,7 +37,7 @@ module.exports = {
 
       await interaction.reply({ content: 'Select a watch-party to remove as watching:', components: [row], ephemeral: true })
 
-      const filter = async (i: SelectMenuInteraction) => {
+      const filter = async (i: any) => {
         if (i.user.id === interaction.user.id) {
           const titleToSet = i.values[0]
           thisServerParty.server.current = thisServerParty.server.current.filter(x => x.title !== titleToSet)
@@ -47,7 +46,7 @@ module.exports = {
         }
         return i.user.id === interaction.user.id
       }
-      interaction.channel!.awaitMessageComponent({ filter, componentType: 'SELECT_MENU', time: 15000 })
+      interaction.channel!.awaitMessageComponent({ filter, componentType: ComponentType.StringSelect, time: 15000 })
     }
   }
 }
