@@ -1,8 +1,7 @@
-import { CommandInteraction } from 'discord.js'
+import { ChatInputCommandInteraction, EmbedBuilder, HexColorString } from 'discord.js'
 import { AniList, AniMedia, AniUser, Parties } from '../../../types'
 
 const { request } = require('graphql-request')
-const Discord = require('discord.js')
 const wait = require('util').promisify(setTimeout)
 const { GET_USERINFO, GET_MEDIALIST, GET_MEDIA } = require('../../../queries')
 
@@ -10,7 +9,7 @@ module.exports = {
   data: {
     name: 'view'
   },
-  async execute (interaction: CommandInteraction, thisServerParty: Parties, _serverAliases: undefined, _serverExists: undefined) {
+  async execute (interaction: ChatInputCommandInteraction, thisServerParty: Parties, _serverAliases: undefined, _serverExists: undefined) {
     const title = interaction.options.getString('title')
     try {
       if (thisServerParty.server.current.length === 0) {
@@ -20,12 +19,12 @@ module.exports = {
         if (thisServerParty.server.current.filter(c => c.title === currentAnime.Media.title.romaji).length > 0) {
           const currentId = currentAnime.Media.id
           interaction.deferReply()
-          const embed = new Discord.MessageEmbed()
-            .setColor(currentAnime.Media.coverImage.color)
+          const embed = new EmbedBuilder()
+            .setColor(currentAnime.Media.coverImage.color as HexColorString)
             .setTitle('Watch Party')
             .setDescription(`Progress on [**${currentAnime.Media.title.romaji}**](${currentAnime.Media.siteUrl})`)
             .setThumbnail(currentAnime.Media.coverImage.large)
-            .setFooter(`requested by ${interaction.user.username}`, `https://cdn.discordapp.com/avatars/${interaction.user.id}/${interaction.user.avatar}.png`)
+            .setFooter({ text: `requested by ${interaction.user.username}`, iconURL: `https://cdn.discordapp.com/avatars/${interaction.user.id}/${interaction.user.avatar}.png` })
             .setTimestamp()
 
           thisServerParty.server.list.find(x => x.animeId === currentId)!.members.forEach(async x => {
@@ -33,10 +32,10 @@ module.exports = {
             try {
               const list: AniList = await request('https://graphql.anilist.co', GET_MEDIALIST, { userName: user.User.name, mediaId: currentAnime!.Media.id })
               const episodes = list.MediaList.progress
-              embed.addField(user.User.name, `[${episodes}/${currentAnime!.Media.episodes}](${user.User.siteUrl})`, true)
+              embed.addFields({ name: user.User.name, value: `[${episodes}/${currentAnime!.Media.episodes}](${user.User.siteUrl})`, inline: true })
             } catch {
               const episodes = 0
-              embed.addField(user.User.name, `[${episodes}/${currentAnime!.Media.episodes}](${user.User.siteUrl})`, true)
+              embed.addFields({ name: user.User.name, value: `[${episodes}/${currentAnime!.Media.episodes}](${user.User.siteUrl})`, inline: true })
             }
           })
           await wait(1000)
